Display recipe cards in main list view

+++ app/src/main/java/com/example/mahima/yummly/RecipeIngredient.java
@@ -5,13 +5,13 @@ import com.google.gson.annotations.SerializedName;
 public class RecipeIngredient {
 
     @SerializedName("quantity")
-    private Integer quantity;
+    private double quantity;
     @SerializedName("measure")
     private String measure;
     @SerializedName("ingredient")
     private String ingredient;
 
-    public Integer getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
